* Converting type of 'MapContract.ContinentRectangle' to 'double[][]' * Converting type of 'MapContract.MapRectangle' to 'double[][]'
git-tfs-id: [https://tfs.codeplex.com:443/tfs/TFS35]$/gw2dotnet/GW2.NET;C37292

+++ GW2.NET/V1/Maps.Contracts/MapContract.cs
@@ -25,9 +25,8 @@ namespace GW2DotNET.V1.Maps.Contracts
         public string ContinentName { get; set; }
 
         /// <summary>Gets or sets the dimensions of the map within the continent coordinate system.</summary>
-        /// TODO: convert to 'double[][]'
         [DataMember(Name = "continent_rect", Order = 10)]
-        public int[][] ContinentRectangle { get; set; }
+        public double[][] ContinentRectangle { get; set; }
 
         /// <summary>Gets or sets the default floor of this map.</summary>
         [DataMember(Name = "default_floor", Order = 3)]
@@ -42,9 +41,8 @@ namespace GW2DotNET.V1.Maps.Contracts
         public string MapName { get; set; }
 
         /// <summary>Gets or sets the dimensions of the map.</summary>
-        /// TODO: convert to 'double[][]'
         [DataMember(Name = "map_rect", Order = 9)]
-        public int[][] MapRectangle { get; set; }
+        public double[][] MapRectangle { get; set; }
 
         /// <summary>Gets or sets the maximum level of this map.</summary>
         [DataMember(Name = "max_level", Order = 2)]
